주문 생성

+++ src/main/java/com/example/demo/domain/Product.java
@@ -23,5 +23,5 @@ public class Product {
     private int quantity;
 
     @Column(nullable=false)
-    private double price;
+    private int price;
 }
